==> Notebooks/1_xr_Process_wav_to_netcdf.ipynb <==
import numpy as np
import pandas as pd
import math
import time
import re
import os

from skimage import util
from scipy.io import wavfile
from scipy import signal
from scipy import stats

import xarray as xr

%matplotlib inline
import matplotlib.pyplot as plt

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

import warnings; warnings.simplefilter('ignore')
np.set_printoptions(suppress=True)

def create_slice_from_wav(file_path, file_len, slice_len, step_size):
    """Creates small slices from wav file. file_length is desired length of file in minutes.
    slice_len is desired legth of each slice in ms. 
    step_size is how big of step to take between steps (larger size is less overlap)."""
    
    #read in wav file
    samp_freq, sig_data = wavfile.read(file_path)
    print('Sampling frequency: ' + str(samp_freq))
    
    #convert file_length from min to sample numbers
    file_length_num = file_len*samp_freq*60
    
    #convert slice_len from ms to sample numbers
    slice_sample_num = slice_len*samp_freq/1000
    
    #use only slices within file_length
    sig_data = sig_data[0:file_length_num]
    
    #determine number of samples and length
    n_samples = sig_data.shape[0]
    print('Number of samples: ' + str(n_samples))
    sig_len = n_samples/samp_freq
    print('Length: ' + str(sig_len) + ' sec')
    
    #create slices 
    steps = int(slice_sample_num*step_size)
    slices = util.view_as_windows(sig_data, window_shape=(slice_sample_num,), step=steps)
    print(f'Audio shape: {sig_data.shape}, Sliced audio shape: {slices.shape}')
    
    return samp_freq, sig_data, slices, steps, sig_len

def create_spec_from_slice_array(slices, steps, spec_window, NFFT, samp_freq):
    """Creates fft spectrogram from slice. spec_window is length of each segment (nperseg).
    NFFT is length of the FFT used (nfft). samp_freq is sampling frequency (in Hz) of slice (fs).
    steps is step size between slices"""
    
    spec_slices = {}
    samp_freq_kHz = samp_freq/1000

    for i in range(slices.shape[0]): 
        if i % 5000 == 0:
            print(i)
            
        #spectrogram
        freqs_spec, times, Sx = signal.spectrogram(slices[i,:], fs=samp_freq, nperseg = spec_window, nfft = NFFT)
    
        time_stamp = (i*steps) / samp_freq_kHz
    
        #store as dic
        spec_slices[time_stamp] = Sx

    return spec_slices, freqs_spec, times


def create_xarray_dataset_from_dic(dic, freqs_spec, times):
    """Creates an xarray.Dataset object from a dictionary input."""
    
    slices_combined = {}
    
    for key, fft_slice in dic.items():
        slices_combined[key] = xr.DataArray(fft_slice, dims = ('freq', 'times'), coords = {'freq': freqs_spec, 'times': times})
    slices_Dataset = xr.Dataset(slices_combined).to_array(dim = 'slices')
    
    return slices_Dataset

#testing function
def get_remainders(slices_Dataset, slices, step_size, samp_freq, sig_len):
    slice_remainder = np.ceil((sig_len - (slices.shape[0] * slices.shape[1] * step_size / samp_freq))*1000)
    netcdf_remainder = (sig_len * 1000) - slices_Dataset.slices.values[-1] - 22.5
    
    return slice_remainder, netcdf_remainder

Find path names for each wav file corresponding to annotated data

wav_dir_path = 'C:/Users/Schindler/Documents/ProgrammingFun/USV_python/Data/Wav_files/18.12.07_CPA_pair_3x'

path_names = []
files = os.listdir(wav_dir_path)
for file in files: 
        path_names.append(wav_dir_path + "/" + file)

len(path_names)

file_len=10
slice_len=25
step_size=0.9

spec_window=128
NFFT=512

for path in path_names:
    
    #get animal name from file
    name = re.search("\d\d\d", path).group(0)

    #process wav file of animal corresponding to annotations
    print(str('Begin processing animal # ' + name))

    #create slices
    start = time.time()
    samp_freq, sig_data, slices, steps, sig_len = create_slice_from_wav(path, file_len, slice_len, step_size)
    end = time.time()
    print(str('Slices created in ' + str(end - start) + '  seconds'))

    #create spectrograms
    start = time.time()
    spec_slices, freqs_spec, times = create_spec_from_slice_array(slices, steps, spec_window, NFFT, samp_freq)
    end = time.time()
    print(str('Spectrograms created in ' + str(end - start) + '  seconds'))

    #create xarray Dataset
    start = time.time()
    slices_Dataset = create_xarray_dataset_from_dic(spec_slices, freqs_spec, times)
    end = time.time()
    print(str('xarray created in ' + str(end - start) + '  seconds'))

    #confirm timestamps are correct
    slice_remainder, netcdf_remainder = get_remainders(slices_Dataset, slices, step_size, samp_freq, sig_len)
    if slice_remainder != netcdf_remainder:
        raise Exception('Mismatch between slice and timestamp remainders')
    else:
        #save    
        start = time.time()
        slices_Dataset.to_netcdf(name + '_xr_Dataset.nc')
        end = time.time()
        print(str('xarray saved in ' + str(end - start) + '  seconds'))
